feat: ensure DbDeleteEpisode calls LoadEpisodeByIdRepository with correct id

+++ src/data/episode/delete-episode/db-delete-episode.ts
@@ -1,4 +1,5 @@
 import { DeleteEpisode } from '../../../domain/use-cases/episode/delete-episode';
+import { NotFoundError } from '../../../shared/errors';
 import { DeleteEpisodeRepository } from '../../protocols/episode/delete-episode-repository';
 import { LoadEpisodeByIdRepository } from '../../protocols/episode/load-episode-by-id-repository';
 
@@ -8,6 +9,11 @@ export class DbDeleteEpisode implements DeleteEpisode {
     private readonly loadEpisodeByIdRepository: LoadEpisodeByIdRepository
   ) {}
   async delete(id: number): Promise<void> {
+    const episode = await this.loadEpisodeByIdRepository.loadById(id);
+    if (!episode) {
+      throw new NotFoundError('Episode not found');
+    }
+
     await this.deleteEpisodeRepository.delete(id);
   }
 }
